+++ src/learning/models/knowledge.rs
@@ -5,15 +5,15 @@ pub struct Knowledge {
     pub id: String,
     pub name: Option<String>,
     pub audio: Option<String>,
-    pub level: Option<u32>,
+    pub level: Option<u8>,
 }
 
 pub trait Levelable {
-    fn level(&self) -> u32;
+    fn level(&self) -> u8;
 }
 
 impl Levelable for Knowledge {
-    fn level(&self) -> u32 {
+    fn level(&self) -> u8 {
         self.level.unwrap()
     }
 }
