+++ src/learning/models/level_analytics.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct LevelAnalytics {
-    pub level: u32,
+    pub level: u8,
     pub count: usize,
     pub progress: f32,
 }
